requirements.txt
Django==1.6.5
South==1.0
Whoosh==2.6.0
defusedxml==0.4.1
django-haystack==2.2.0
django-tastypie==0.11.1
lxml==3.3.6
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.7.3
wsgiref==0.1.2
